Run the Boruta algorithm on a sem tree

Grows a series of SEM Forests following the boruta algorithm to determine feature importance as moderators of the underlying model.

boruta(
  model,
  data,
  control = NULL,
  predictors = NULL,
  maxRuns = 30,
  pAdjMethod = "none",
  alpha = 0.05,
  verbose = FALSE,
  quant = 1,
  ...
)

Arguments

model

A template SEM. Same as in semtree.

data

A dataframe to boruta on. Same as in semtree.

control

A semforest control object to set forest parameters.

predictors

An optional list of covariates. See semtree code example.

maxRuns

Maximum number of boruta search cycles

pAdjMethod

A value from p.adjust.methods defining a multiple testing correction method

alpha

p-value cutoff for decision making. Default .05

verbose

Verbosity level for boruta processing similar to the same argument in semtree.control and semforest.control

quant

Quantile for selection. Default 1.

...

Optional parameters to undefined subfunctions

Value

A vim object with several elements that need work. Of particular note, `$importance` carries mean importance; `$decision` denotes Accepted/Rejected/Tentative; `$impHistory` has the entire varimp history; and `$details` has exit values for each parameter.

See also

Author

Priyanka Paul, Timothy R. Brick, Andreas Brandmaier